requirements.txt
pandas
numpy
PyWavelets
librosa
matplotlib
seaborn
plotly
torch
torchvision
torchaudio
wandb
ruamel.yaml
configmypy
zencfg
tensorly
tensorly-torch
torch-harmonics
opt-einsum
h5py
zarr
black
flaky
pytest
neuraloperator